Bivariate copula models

Bivariate copula models

bicop(data, family_set = "all", par_method = "mle",
  nonpar_method = "quadratic", mult = 1, selcrit = "bic", presel = TRUE,
  keep_data = TRUE, cores = 1)

bicop_dist(family = "indep", rotation = 0, parameters = numeric(0))

Arguments

data
a matrix or data.frame (copula data should have approximately uniform margins).

family_set
a character vector of families; see Details for additional options.

par_method
the estimation method for parametric models, either "mle" for maximum likelihood or "itau" for inversion of Kendall's tau (only available for one-parameter families and "t".

nonpar_method
the estimation method for nonparametric models, either "constant" for the standard transformation estimator, or "linear"/"quadratic" for the local-likelihood approximations of order one/two.

mult
multiplier for the smoothing parameters of nonparametric families. Values larger than 1 make the estimate more smooth, values less than 1 less smooth.

selcrit
criterion for family selection, either "loglik", "aic", or "bic".

presel
whether the family set should be thinned out according to symmetry characteristics of the data.

keep_data
whether the data should be stored (necessary for computing fit statistics and using fitted()).

cores
number of cores to use; if more than 1, estimation for multiple families is done in parallel.

family
the copula family, a string containing the family name (see Details for all possible families).

rotation
the rotation of the copula, one of 0, 90, 180, 270.

parameters
a vector or matrix of copula parameters.

Value

Objects inheriting from bicop_dist for bicop_dist(), and bicop and bicop_dist for bicop().

Details

The implemented families are: "indep" = Independence copula. "gaussian" = Gaussian copula. "t" = Student t copula. "clayton" = Clayton copula. "gumbel" = Gumbel copula. "frank" = Frank copula. "joe" = Joe copula. "bb1" = BB1 copula. "bb6" = BB6 copula. "bb7" = BB7 copula. "bb8" = BB8 copula. "tll" = transformation kernel local likelihood, only for bicop().

In addition, the following convenience definitions can be used (and combined) with bicop: "all" = all families. "parametric" = parametric families. "nonparametric" = nonparametric families. "archimedean" = archimedean families. "elliptical" = elliptical families. "bbs" = BB families. "oneparametric" = one parameter families. "twoparametric" = two parameter families. Partial matching is activated. For example, "gauss" is equivalent to "gaussian", or you can write "nonpar" instead of "nonparametric".

Examples

## bicop_dist objects
bicop_dist("gaussian", 0, 0.5)

#> Bivariate copula ('bicop_dist'): family = gaussian, rotation = 0, parameters = 0.5

str(bicop_dist("gauss", 0, 0.5))

#> List of 4
#>  $ family    : chr "gaussian"
#>  $ rotation  : num 0
#>  $ parameters: num [1, 1] 0.5
#>  $ npars     : int 1
#>  - attr(*, "class")= chr "bicop_dist"

bicop <- bicop_dist("clayton", 90, 3)

## bicop objects
u <- rbicop(500, "gauss", 0, 0.5)
fit1 <- bicop(u, "par")
fit1

#> Bivariate copula fit ('bicop'): family = gaussian, rotation = 0, parameters = 0.5195865
#> nobs = 500   logLik = 70.84   npars = 1   AIC = -139.67   BIC = -135.46