High Performance Algorithms for Vine Copula Modeling

Provides an interface to 'vinecopulib', a C++ library for vine copula modeling based on 'Boost' and 'Eigen'. The 'rvinecopulib' package implements the core features of the popular 'VineCopula' package, in particular inference algorithms for both vine copula and bivariate copula models. Advantages over 'VineCopula' are a sleeker and more modern API, improved performances, especially in high dimensions, nonparametric and multi-parameter families. The 'rvinecopulib' package includes 'vinecopulib' as header-only C++ library (currently version 0.2.4). Thus users do not need to install 'vinecopulib' itself in order to use 'rvinecopulib'. Since their initial releases, 'vinecopulib' is licensed under the MIT License, and 'rvinecopulib' is licensed under the GNU GPL version 3.

Examples

## bicop_dist objects
bicop_dist("gaussian", 0, 0.5)

#> Bivariate copula ('bicop_dist'): family = gaussian, rotation = 0, parameters = 0.5

str(bicop_dist("gauss", 0, 0.5))

#> List of 4
#>  $ family    : chr "gaussian"
#>  $ rotation  : num 0
#>  $ parameters: num [1, 1] 0.5
#>  $ npars     : int 1
#>  - attr(*, "class")= chr "bicop_dist"

bicop <- bicop_dist("clayton", 90, 3)

## bicop objects
u <- rbicop(500, "gauss", 0, 0.5)
fit1 <- bicop(u, "par")
fit1

#> Bivariate copula fit ('bicop'): family = gaussian, rotation = 0, parameters = 0.51
#> nobs = 500   logLik = 82.95   npars = 1   AIC = -163.89   BIC = -159.68   

## vinecop_dist objects
## specify pair-copulas
bicop <- bicop_dist("bb1", 90, c(3, 2))
pcs <- list(
    list(bicop, bicop),  # pair-copulas in first tree 
    list(bicop)          # pair-copulas in second tree 
)
## specify R-vine matrix
mat <- matrix(c(1, 2, 3, 1, 2, 0, 1, 0, 0), 3, 3)
## build the vinecop_dist object
vc <- vinecop_dist(pcs, mat)
summary(vc)

#> 3-dimensional vine copula model
#> -------------------------------
#>  tree edge conditioned conditioning family rotation parameters
#>     1    1        3, 1                 bb1       90       3, 2
#>     1    2        2, 1                 bb1       90       3, 2
#>     2    1        3, 2            1    bb1       90       3, 2

## vinecop objects
u <- sapply(1:3, function(i) runif(50))
vc <- vinecop(u, "par")
summary(vc)

#> 3-dimensional vine copula model
#>  fit
#> nobs = 50   logLik = 0   npars = 0   AIC = 0   BIC = 0   
#> ----
#> 3-dimensional vine copula model fitnobs = 50   logLik = 0   npars = 0   AIC = 0   BIC = 0   
#> --------------------------------------------------------------------------------------------
#>  tree edge conditioned conditioning family rotation parameters
#>     1    1        2, 1               indep        0           
#>     1    2        1, 3               indep        0           
#>     2    1        2, 3            1  indep        0