Fit vine copula models

Automated fitting for vine copula models

vinecop(data, family_set = "all", matrix = NA, par_method = "mle",
  nonpar_method = "constant", mult = 1, selcrit = "bic", presel = TRUE,
  trunc_lvl = Inf, tree_crit = "tau", threshold = 0, keep_data = TRUE)

Arguments

data
a matrix or data.frame (copula data should have approximately uniform margins).

family_set
a character vector of families; as in bicop_dist(), see Details for additional options.

matrix
an R-vine matrix specifying the structure matrix (see check_rvine_matrix()), or NA for automatic structure selection (default).

par_method
the estimation method for parametric models, either "mle" for maximum likelihood or "itau" for inversion of Kendall's tau (only available for one-parameter families and "t".

nonpar_method
the estimation method for nonparametric models, either "constant" for the standard transformation estimator, or "linear"/"quadratic" for the local-likelihood approximations of order one/two.

mult
multiplier for the smoothing parameters of nonparametric families. Values larger than 1 make the estimate more smooth, values less than 1 less smooth.

selcrit
criterion for family selection, either "loglik", "aic", or "bic".

presel
whether the family set should be thinned out according to symmetry characteristics of the data.

trunc_lvl
the truncation level of the vine copula.

tree_crit
the criterion for tree selection, one of "tau", "rho", "hoeffd" for Kendall's \(tau\), Spearman's \(rho\), and Hoeffding's \(D\), respectively.

threshold
for thresholded vine copulas.

keep_data
whether the data should be stored (necessary for computing fit statistics and using fitted()).

Value

An object inherting from bicop and bicop_dist.

Details

In addition, to the families in bicop_dist(), the following convenience defintion can be used (and combined):

"all"

all families

.
"parametric"

parametric families.

"nonparametric"

nonparametric families.

"archimedean"

archimedean families.

"elliptical"

elliptical families.

"bbs"

BB families.

"oneparametric "

one parameter families.

"twoparametric "

two parameter families.

Partial matching is activated. For example, you can write "nonpar" instead of the full name "nonparametric".

Examples

u <- sapply(1:3, function(i) runif(50))
fit <- vinecop(u, "par")
fit

#> 3-dimensional vine copula model fit
#> nobs = 50   logLik = 0   npars = 0   AIC = 0   BIC = 0   

summary(fit)

#> 3-dimensional vine copula model fit
#> nobs = 50   logLik = 0   npars = 0   AIC = 0   BIC = 0   
#> ----
#>  tree edge conditioned conditioning family rotation parameters
#>     1    1        2, 3               indep        0           
#>     1    2        3, 1               indep        0           
#>     2    1        2, 1            3  indep        0           

str(fit, 3)

#> List of 5
#>  $ pair_copulas:List of 2
#>   ..$ :List of 2
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>   ..$ :List of 1
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>  $ matrix      : num [1:3, 1:3] 3 1 2 1 3 0 1 0 0
#>  $ data        : num [1:50, 1:3] 0.334 0.749 0.913 0.204 0.472 ...
#>  $ controls    :List of 9
#>   ..$ family_set   : chr [1:11] "indep" "gaussian" "clayton" "gumbel" ...
#>   ..$ par_method   : chr "mle"
#>   ..$ nonpar_method: chr "constant"
#>   ..$ mult         : num 1
#>   ..$ selcrit      : chr "bic"
#>   ..$ presel       : logi TRUE
#>   ..$ trunc_lvl    : num Inf
#>   ..$ tree_crit    : chr "tau"
#>   ..$ threshold    : num 0
#>  $ nobs        : int 50
#>  - attr(*, "class")= chr [1:2] "vinecop" "vinecop_dist"