Vine copula models

Automated fitting or creation of custom vine copula models

vinecop(data, family_set = "all", matrix = NA, par_method = "mle",
  nonpar_method = "constant", mult = 1, selcrit = "bic", presel = TRUE,
  trunc_lvl = Inf, tree_crit = "tau", threshold = 0, keep_data = TRUE,
  show_trace = FALSE)

vinecop_dist(pair_copulas, matrix)

Arguments

data
a matrix or data.frame (copula data should have approximately uniform margins).

family_set
a character vector of families; see bicop() for additional options.

matrix
a quadratic matrix specifying the structure matrix (see check_rvine_matrix()); for vinecop_dist(), the dimension must be length(pair_copulas)-1; for vinecop(), matrix = NA performs automatic structure selection.

par_method
the estimation method for parametric models, either "mle" for maximum likelihood or "itau" for inversion of Kendall's tau (only available for one-parameter families and "t".

nonpar_method
the estimation method for nonparametric models, either "constant" for the standard transformation estimator, or "linear"/"quadratic" for the local-likelihood approximations of order one/two.

mult
multiplier for the smoothing parameters of nonparametric families. Values larger than 1 make the estimate more smooth, values less than 1 less smooth.

selcrit
criterion for family selection, either "loglik", "aic", or "bic".

presel
whether the family set should be thinned out according to symmetry characteristics of the data.

trunc_lvl
the truncation level of the vine copula; Inf means no truncation, NA indicates that the truncation level should be selected automatically by GIC.

tree_crit
the criterion for tree selection, one of "tau", "rho", "hoeffd" for Kendall's \(tau\), Spearman's \(rho\), and Hoeffding's \(D\), respectively.

threshold
for thresholded vine copulas; NA indicates that the threshold should be selected automatically by GIC.

keep_data
whether the data should be stored (necessary for computing fit statistics and using fitted()).

show_trace
logical; whether a trace of the fitting progress should be printed.

pair_copulas
A nested list of 'bicop_dist' objects, where pair_copulas[[t]][[e]] corresponds to the pair-copula at edge e in tree t.

Value

Objects inheriting from vinecop_dist for vinecop_dist(), and vinecop and vinecop_dist for vinecop().

Details

vinecop_dist() creates a vine copula by specifying a nested list of bicop_dist objects and a quadratic structure matrix. vinecop() provides automated fitting for vine copula models. The function inherits the parameters of bicop(). Optionally, a quadratic matrix can be used as input to pre-specify the vine structure. tree_crit describes the criterion for tree selection, one of "tau", "rho", "hoeffd" for Kendall's tau, Spearman's rho, and Hoeffding's D, respectively. Additionally, threshold allows to threshold the tree_crit and trunc_lvl to truncate the vine copula, with threshold_sel and trunc_lvl_sel to automatically select both parameters.

Examples

# specify pair-copulas
bicop <- bicop_dist("bb1", 90, c(3, 2))
pcs <- list(
    list(bicop, bicop),  # pair-copulas in first tree 
    list(bicop)          # pair-copulas in second tree 
 )

# specify R-vine matrix
mat <- matrix(c(1, 2, 3, 1, 2, 0, 1, 0, 0), 3, 3)

# set up vine copula model
vc <- vinecop_dist(pcs, mat)

# show model
summary(vc)

#>  tree edge conditioned conditioning family rotation parameters
#>     1    1        3, 1                 bb1       90       3, 2
#>     1    2        2, 1                 bb1       90       3, 2
#>     2    1        3, 2            1    bb1       90       3, 2

# get some data
u <- sapply(1:3, function(i) runif(50))

# estimate a vine copula model
fit <- vinecop(u, "par")
fit

#> 3-dimensional vine copula model fit
#> nobs = 50   logLik = 2.38   npars = 1   AIC = -2.76   BIC = -0.85   

summary(fit)

#> 3-dimensional vine copula model fit
#> nobs = 50   logLik = 2.38   npars = 1   AIC = -2.76   BIC = -0.85   
#> ----
#>  tree edge conditioned conditioning  family rotation parameters
#>     1    1        3, 1              clayton      270        0.4
#>     1    2        2, 1                indep        0           
#>     2    1        3, 2            1   indep        0           

str(fit, 3)

#> List of 6
#>  $ pair_copulas:List of 2
#>   ..$ :List of 2
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>   ..$ :List of 1
#>   .. ..$ :List of 4
#>   .. .. ..- attr(*, "class")= chr [1:2] "bicop" "bicop_dist"
#>  $ matrix      : num [1:3, 1:3] 1 2 3 1 2 0 1 0 0
#>  $ npars       : num 1
#>  $ data        : num [1:50, 1:3] 0.175 0.134 0.395 0.328 0.614 ...
#>  $ controls    :List of 9
#>   ..$ family_set   : chr [1:11] "indep" "gaussian" "clayton" "gumbel" ...
#>   ..$ par_method   : chr "mle"
#>   ..$ nonpar_method: chr "constant"
#>   ..$ mult         : num 1
#>   ..$ selcrit      : chr "bic"
#>   ..$ presel       : logi TRUE
#>   ..$ trunc_lvl    : num Inf
#>   ..$ tree_crit    : chr "tau"
#>   ..$ threshold    : num 0
#>  $ nobs        : int 50
#>  - attr(*, "class")= chr [1:2] "vinecop" "vinecop_dist"